# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Private Practice Management & Client Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks (filtered by role/assignment)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, in_progress, completed, overdue]
        - in: query
          name: assigneeId
          schema:
            type: string
        - in: query
          name: caseId
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create and assign a task (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a single task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task complete (with optional delay log)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCompletion"
      responses:
        "200":
          description: Completed task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "422":
          description: Delay reason required

  # ── Delay Logs ─────────────────────────────────────────────────────────────
  /tasks/{id}/delay-log:
    post:
      operationId: createDelayLog
      tags: [delay-logs]
      summary: Submit a delay reason for an overdue task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DelayLogInput"
      responses:
        "201":
          description: Delay log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelayLog"

  # ── Consultations ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, caseId, title, status, priority, deadline, createdAt, updatedAt]
      properties:
        id:
          type: integer
        caseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, overdue]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          type: ["string", "null"]
        assigneeName:
          type: ["string", "null"]
        deadline:
          type: string
          format: date
        completedAt:
          type: ["string", "null"]
          format: date-time
        aiAllowed:
          type: boolean
          description: >-
            Whether the assignee may use AI drafting on this task's matter. The
            per-task grant: a junior advocate holds `drafting.use` but is not
            trusted with the chamber's whole AI budget, so whoever assigns the
            work decides task by task. Admin and senior advocate are
            chamber-wide and never consult it; a clerk never holds
            `drafting.use`, so it is inert for them. Settable only by a caller
            with `tasks.write`.
        isOverdue:
          type: boolean
        hasDelayLog:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [caseId, title, deadline]
      properties:
        caseId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        assigneeId:
          type: string
        deadline:
          type: string
          format: date
        aiAllowed:
          type: boolean
          default: false
          description: Grant AI drafting on this task's matter to its assignee.
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, overdue]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          type: string
        deadline:
          type: string
          format: date
        aiAllowed:
          type: boolean
          description: Grant or withdraw AI drafting on this task's matter.
    TaskCompletion:
      type: object
      properties:
        delayReason:
          type: string
          enum: [client_unresponsive, court_delay, document_missing, resource_unavailable, other]
        delayNotes:
          type: string
        proofFileName:
          type: string

    # ── Delay Log ─────────────────────────────────────────────────────────
    DelayLog:
      type: object
      required: [id, taskId, reason, submittedAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        reason:
          type: string
          enum: [client_unresponsive, court_delay, document_missing, resource_unavailable, other]
        notes:
          type: ["string", "null"]
        proofFileName:
          type: ["string", "null"]
        submittedAt:
          type: string
          format: date-time
    DelayLogInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum: [client_unresponsive, court_delay, document_missing, resource_unavailable, other]
        notes:
          type: string
        proofFileName:
          type: string

    # ── Consultation ──────────────────────────────────────────────────────
...